Add 1px solid #e0e0e0 border to toc
without *some* border, the toc doesn't look good,
is too radical of a change.

with the 1px border, looks ok.

this is also going to work better with the way
we arrange the TOC on wikidata (although suppose
we could provide a css override in wikibase)

http://snag.gy/Pr5Le.jpg (before)

http://snag.gy/JniR8.jpg (after)

(maybe at a later time, further change to remove
the border again might make sense but not yet
and maybe not at all)

Change-Id: I2fef239259fe7048944e06446703e4574d827f71